Extract linting from tasks.py
